refacotr: CropScaler
